+++ packages/rust/napi-rs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vrn-deco/boilerplate-rust-napi-rs
 
+## 1.0.2
+
+### Patch Changes
+
+- ESMify
+- Updated dependencies
+  - @vrn-deco/boilerplate-preset-base@1.3.3
+
 ## 1.0.1
 
 ### Patch Changes