+++ packages/typescript/vue3-vant-h5plus-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vrn-deco/boilerplate-typescript-vue3-vant-h5plus-app
 
+## 1.1.19
+
+### Patch Changes
+
+- ESMify
+- Updated dependencies
+  - @vrn-deco/boilerplate-preset-npm@1.1.4
+
 ## 1.1.18
 
 ### Patch Changes